+++ person.rb
@@ -1,6 +1,6 @@
-require_relative 'name_able'
-require_relative 'trimmer_decorator '
-require_relative 'capital_decorator'
+require './name_able'
+require './trimmer_decorator '
+require './capital_decorator'
 require './rental'
 
 class Person < Nameable
